fix(server): keys 계산 로직 오류 수정

+++ core/src/main/kotlin/com/devjk/djtrinity/domain/BmsHeader.kt
@@ -27,7 +27,7 @@ class BmsHeader private constructor(
     var random: Double? = null,
 
     // custom values
-    var keys: Double? = null,
+    var keys: Int? = null,
 
     header: String,
     parseWav: Boolean
